Use adjusted relatedTarget on WindowEventContext
In event capturing phase, when the event is dispatched on window, the
related target was not set properly (it uses the state of the event
at the target, which means related target is the same as when it was
at the target). This might cause leaks when the event's relatedTarget
is inside a shadow tree. This change makes it so at window context,
the relatedTarget used is the relatedTarget of the last
NodeEventContext.

Bug: 817222
Change-Id: I1edea3977202b0a05e3e8ea46a5b8c1edb4b834d